requirements.txt
streamlit
requests
Pillow
psycopg2-binary
boto3
botocore
python-dotenv
